md_ops: cache the validated chain leading up to a merkle leaf
When verifying the MDs written by a revoked device, we have to
validate a long chain of MDs between the MD we're looking at, and the
MD just following the revocation.  if we're doing something like
garbage collection and iterating through MDs, we may have to do this
repeatedly for the "end" MD in the chain.  If that chain exceeds the
size of the in-memory MD cache, we will be going to the mdserver a lot
to download the same MDs over and over again.

Instead, introduce a simple cache that maps the "end" MD to the
earlier revision in a validated chain.  That way we can avoid all the
extra downloads when we miss the cache, and just download the MDs we
care about.

Issue: KBFS-3581